requirement.txt
Flask==2.3.3
flask-cors==4.0.0

python-docx==1.1.0
PyPDF2==3.0.1

googletrans==4.0.0rc1

openai==1.30.1

gunicorn==21.2.0  # Optional: if deploying to production (e.g., Heroku)

